Se eliminan tablas Mecanic y CapTaller ya que se emplean los datos de usuario directamente.
Se vincula mediante clave foranea los campos idmecanic e idcaptaller de la tabla Reparacio al campo id de la tabla Usuario

+++ pac4/server/src/main/java/ss3/dao/impl/ReparacionDAOImpl.java
@@ -562,7 +562,7 @@ public class ReparacionDAOImpl extends GenericDaoImpl implements ReparacionDAO {
             ptmt.setBoolean(2, rep.isAceptada());
             ptmt.setInt(3, rep.getIdMecanico());
             ptmt.setBoolean(4, rep.isAsignada());
-            ptmt.setDouble(5, rep.getContador());
+            ptmt.setInt(5, rep.getContador());
             ptmt.setString(6, rep.getObservaciones());
             ptmt.setInt(7,rep.getNumcom());
             if (ptmt.executeUpdate() > 0) {
